+++ GA-AppLocker/GA-AppLocker.psm1
@@ -417,6 +417,7 @@ Export-ModuleMember -Function @(
     'New-DeploymentJob',
     'Get-DeploymentJob',
     'Get-AllDeploymentJobs',
+    'Remove-DeploymentJob',
     'Start-Deployment',
     'Stop-Deployment',
     'Get-DeploymentStatus',