+++ GA-AppLocker/Modules/GA-AppLocker.Deployment/GA-AppLocker.Deployment.psm1
@@ -73,6 +73,7 @@ Export-ModuleMember -Function @(
     'New-DeploymentJob',
     'Get-DeploymentJob',
     'Get-AllDeploymentJobs',
+    'Remove-DeploymentJob',
     'Start-Deployment',
     'Stop-Deployment',
     'Get-DeploymentStatus',